Renaming the Assign class to LableCreate. Changing the syntax accordingly.
