+++ label-manager/createlabel/src/main/java/se/tre/freki/labelmanager/CreateLabel.java
@@ -1,4 +1,4 @@
-package se.tre.freki.idmanager;
+package se.tre.freki.labelmanager;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -28,27 +28,27 @@ import javax.annotation.Nullable;
 import javax.inject.Inject;
 
 /**
- * Command line tool to assign new IDs.
+ * Command line tool to create new Label IDs.
  */
-public final class Assign {
-  private static final Logger LOG = LoggerFactory.getLogger(Assign.class);
+public final class CreateLabel {
+  private static final Logger LOG = LoggerFactory.getLogger(CreateLabel.class);
 
   private final LabelClient labelClient;
 
   @Inject
-  Assign(final LabelClient labelClient) {
+  CreateLabel(final LabelClient labelClient) {
     this.labelClient = checkNotNull(labelClient);
   }
 
   /**
-   * Entry-point for the assign application. The assign program is normally not executed directly
-   * but rather through the main project.
+   * Entry-point for the createLabel application. The createLabel program is normally not
+   * executed directly but rather through the main project.
    *
    * @param args The command-line arguments
    */
   public static void main(final String[] args) {
     final CommandLineApplication application = CommandLineApplication.builder()
-        .command("id assign")
+        .command("label create")
         .usage("[OPTIONS] <TYPE> [NAME]...")
         .description("Create IDs for NAME(s), or names read from standard input of type TYPE.")
         .helpText("With no NAME, or when NAME is -, read standard input.")
@@ -65,7 +65,7 @@ public final class Assign {
 
       cmdOptions.configureLogger();
 
-      final AssignComponent assignComponent = DaggerAssignComponent.builder()
+      final CreateLabelComponent createLabelComponent = DaggerCreateLabelComponent.builder()
           .configModule(cmdOptions.configModule())
           .build();
 
@@ -75,14 +75,14 @@ public final class Assign {
       final ImmutableSet<String> names = ImmutableSet.copyOf(
           Arrays.copyOfRange(args, 1, args.length));
 
-      final Store store = assignComponent.store();
-      final Assign assign = assignComponent.assign();
+      final Store store = createLabelComponent.store();
+      final CreateLabel createLabel = createLabelComponent.createLabel();
 
       final List<ListenableFuture<LabelId>> assignments =
           Lists.newArrayListWithCapacity(names.size());
 
       for (final String name : names) {
-        assignments.add(assign.assign(name, type));
+        assignments.add(createLabel.createLabel(name, type));
       }
 
       Futures.allAsList(assignments).get();
@@ -94,7 +94,7 @@ public final class Assign {
       System.err.println(e.getMessage());
       System.exit(42);
     } catch (Exception e) {
-      LOG.error("Fatal error while assigning id", e);
+      LOG.error("Fatal error while creating id", e);
       System.exit(42);
     }
   }
@@ -104,11 +104,11 @@ public final class Assign {
       String stringType = nonOptionArguments.get(0).toString();
       return LabelType.fromValue(stringType);
     } catch (IndexOutOfBoundsException e) {
-      throw new IllegalArgumentException("Missing identifier type to assign");
+      throw new IllegalArgumentException("Missing identifier type to create label");
     }
   }
 
-  private ListenableFuture<LabelId> assign(final String name, final LabelType type) {
+  private ListenableFuture<LabelId> createLabel(final String name, final LabelType type) {
     final ListenableFuture<LabelId> id = labelClient.createId(type, name);
     Futures.addCallback(id, new LogNewIdCallback(name, type));
     return id;
